Added Cloudinary image upload, seller details and admin improvements

+++ backend/marketplace-backend/src/main/java/com/communitymarketplace/marketplacebackend/service/AdminService.java
@@ -33,6 +33,13 @@ public class AdminService {
                 .orElseThrow(() ->
                         new RuntimeException("User not found"));
 
+        if(user.getRole().equalsIgnoreCase("ADMIN")) {
+
+            throw new RuntimeException(
+                    "Admin cannot be deleted"
+            );
+        }
+
         userRepository.delete(user);
 
         return "User deleted successfully";
